First round of work towards federation: record deleted and lastmoddate on libs and lib items
Squashed commit of the following:

commit 65d6cc3841a8d0f1e89f2fb687a89324ff270a72
Author: drlippman <drlippman@yahoo.com>
Date:   Fri Jun 23 17:25:08 2017 -0700

    Disable use of federatedapi for now

commit 80c7fa7317512c308185417a8f43041eed73129c
Merge: 7aa697b aff9a57
Author: drlippman <drlippman@yahoo.com>
Date:   Fri Jun 23 17:20:09 2017 -0700

    Merge branch 'feature.federation' of github.com:drlippman/IMathAS into feature.federation

    Conflicts:
    	admin/importitems.php

commit 7aa697b35cbf27cc41460c61c09d916ef9148bbf
Author: drlippman <drlippman@yahoo.com>
Date:   Fri Jun 23 17:14:35 2017 -0700

    Minor bug fix importitems

commit 645bdc5ad68f29e8abaad950da07ed859651c9df
Author: drlippman <drlippman@yahoo.com>
Date:   Tue Jun 13 20:08:42 2017 -0700

    Bug fix lib item deleted handling

commit f85838481f72965a794a8a567d9db02d45b7c46f
Author: drlippman <drlippman@yahoo.com>
Date:   Mon Jun 12 23:18:12 2017 -0700

    Fix issues from rebase

commit 799584d25600d091f1db10b3248b1cafea02316e
Author: drlippman <drlippman@yahoo.com>
Date:   Thu Feb 9 19:01:15 2017 -0800

    Start work on federation pull

commit fe656e71a0f0af25131a4084114d0b252892a54f
Author: drlippman <drlippman@yahoo.com>
Date:   Tue Feb 7 19:50:33 2017 -0800

    Initial version of federatedapi endpoint

commit fa2a38cacad5d255a9df4b870c107bcfd05cf99f
Author: drlippman <drlippman@yahoo.com>
Date:   Mon Feb 6 17:38:45 2017 -0800

    Revamp qimages and instr_files to allow urls
    Rework item import/export to always use URLs for files

commit 8ee30f3248ed00776c2cd27ce7a67870c6a77b1d
Author: drlippman <drlippman@yahoo.com>
Date:   Mon Feb 6 16:38:10 2017 -0800

    Adjust exportitems/importitems for deleted

commit f3c6ec5ddeb70e2156bcbc9768262ca9f6672e51
Author: drlippman <drlippman@yahoo.com>
Date:   Mon Feb 6 14:39:53 2017 -0800

    Adjust exportlib/importlib for deleted

commit 0cd9d63a8ab1ae19c5f65c634cf6fc55a89bc941
Author: drlippman <drlippman@yahoo.com>
Date:   Mon Feb 6 13:28:44 2017 -0800

    Remove questionset import/export; redundant with other methods

commit fcc9ba78ba2b52982f1da1f3781f38b829270052
Author: drlippman <drlippman@yahoo.com>
Date:   Sun Feb 5 20:27:16 2017 -0800

    Bug fix managelibs libs deleted

commit 4b8f7aba83db32628b8a42dd889ce4f2f08a5324
Author: drlippman <drlippman@yahoo.com>
Date:   Sun Feb 5 18:20:54 2017 -0800

    Bug fix deleted lib items handling manageqset

commit 97599c6efdbc8853d82c9ecc8e790f5e1ac6d688
Author: drlippman <drlippman@yahoo.com>
Date:   Sun Feb 5 17:11:57 2017 -0800

    bug fix moddataset library handling

commit 99e12ea75544989f4608b817c69a1a28658d4cc5
Author: drlippman <drlippman@yahoo.com>
Date:   Sun Feb 5 16:26:46 2017 -0800

    Update primary pages to use deleted for libs, lib items

commit bc93fb23b3d84a30cb0a10a6d452c49ad624fb35
Author: drlippman <drlippman@yahoo.com>
Date:   Sat Feb 4 15:44:30 2017 -0800

    Add first set of simple screens for deleted libraries/lib items

commit 5f88dcefdc907efd3728f9c8e3aa46ddd27311fd
Author: drlippman <drlippman@yahoo.com>
Date:   Thu Feb 2 21:36:46 2017 -0800

    Add page for add/modify fed peers

commit aff9a575273e3cf2847d5cc332ac173b77121b29
Author: David Lippman <drlippman@yahoo.com>
Date:   Tue Jun 13 23:48:24 2017 -0700

    Update lastmoddate when junkflag set in lib items

commit dff6674419e2fcd5ba19c0cc4c9d1323da52e507
Author: drlippman <drlippman@yahoo.com>
Date:   Tue Jun 13 20:08:42 2017 -0700

    Bug fix lib item deleted handling

commit a696f95654f4b00918e2d8207d20244e98b0d825
Author: drlippman <drlippman@yahoo.com>
Date:   Mon Jun 12 23:18:12 2017 -0700

    Fix issues from rebase

commit 71dc2722b4adb509165192d42a820bf2cd2360a1
Author: drlippman <drlippman@yahoo.com>
Date:   Thu Feb 9 19:01:15 2017 -0800

    Start work on federation pull

commit e8701fee53ebe098e65dacc6a6aa8435c3f4bed5
Author: drlippman <drlippman@yahoo.com>
Date:   Tue Feb 7 19:50:33 2017 -0800

    Initial version of federatedapi endpoint

commit dc4bb8acf213b079ddd8d82f8a579b63f7a89eee
Author: drlippman <drlippman@yahoo.com>
Date:   Mon Feb 6 17:38:45 2017 -0800

    Revamp qimages and instr_files to allow urls
    Rework item import/export to always use URLs for files

commit 5997e189b5c61a9e6343c297b049dcfe31dc8b5c
Author: drlippman <drlippman@yahoo.com>
Date:   Mon Feb 6 16:38:10 2017 -0800

    Adjust exportitems/importitems for deleted

commit c8088a5d74ce3661ba1d00153820518cb291bc69
Author: drlippman <drlippman@yahoo.com>
Date:   Mon Feb 6 14:39:53 2017 -0800

    Adjust exportlib/importlib for deleted

commit 798ad1f343232033f43f9889a0d8fb478b5f9f59
Author: drlippman <drlippman@yahoo.com>
Date:   Mon Feb 6 13:28:44 2017 -0800

    Remove questionset import/export; redundant with other methods

commit 3f372981539cf3e254d6e3bdaeec93ab3bf258fa
Author: drlippman <drlippman@yahoo.com>
Date:   Sun Feb 5 20:27:16 2017 -0800

    Bug fix managelibs libs deleted

commit f3b5e2dc47f633319f9512b668ff1979aaa3e365
Author: drlippman <drlippman@yahoo.com>
Date:   Sun Feb 5 18:20:54 2017 -0800

    Bug fix deleted lib items handling manageqset

commit 6625a4e50ad79b14ece9e2ddd701e2529e388b37
Author: drlippman <drlippman@yahoo.com>
Date:   Sun Feb 5 17:11:57 2017 -0800

    bug fix moddataset library handling

commit 67ac5016757ded69da13e69c0c7ec0c792962e6e
Author: drlippman <drlippman@yahoo.com>
Date:   Sun Feb 5 16:26:46 2017 -0800

    Update primary pages to use deleted for libs, lib items

commit 679bd89124c280103a24c379a8bea3fd06954160
Author: drlippman <drlippman@yahoo.com>
Date:   Sat Feb 4 15:44:30 2017 -0800

    Add first set of simple screens for deleted libraries/lib items

commit 21e0079452cee97349556d710eaf25df9ef215b1
Author: drlippman <drlippman@yahoo.com>
Date:   Thu Feb 2 21:36:46 2017 -0800

    Add page for add/modify fed peers

+++ course/reviewlibrary.php
@@ -79,7 +79,7 @@ if ($myrights<20) {
 		//DB $query = "SELECT count(qsetid) FROM imas_library_items WHERE libid='$lib'";
 		//DB $result = mysql_query($query) or die("Query failed : $query " . mysql_error());
 		//DB $cnt = mysql_result($result,0,0);
-		$stm = $DBH->prepare("SELECT count(qsetid) FROM imas_library_items WHERE libid=:libid");
+		$stm = $DBH->prepare("SELECT count(qsetid) FROM imas_library_items WHERE libid=:libid AND deleted=0");
 		$stm->execute(array(':libid'=>$lib));
 		$cnt = $stm->fetchColumn(0);
 		if ($cnt==0) {
@@ -91,7 +91,7 @@ if ($myrights<20) {
 		//DB $result = mysql_query($query) or die("Query failed : $query " . mysql_error());
 		//DB $qsetid = mysql_result($result,0,0);
 		$offset = intval($offset);
-		$stm = $DBH->prepare("SELECT qsetid FROM imas_library_items WHERE libid=:libid LIMIT $offset,1");
+		$stm = $DBH->prepare("SELECT qsetid FROM imas_library_items WHERE libid=:libid AND deleted=0 LIMIT $offset,1");
 		$stm->execute(array(':libid'=>$lib));
 		$qsetid = $stm->fetchColumn(0);
 
@@ -174,7 +174,8 @@ if ($myrights<20) {
 						//DB if (mysql_num_rows($result)>0) {
 						$query = "SELECT imas_library_items FROM imas_library_items,imas_users WHERE ";
 						$query .= "imas_library_items.ownerid=imas_users.id AND imas_users.groupid=:groupid AND ";
-						$query .= "imas_library_items.qsetid=:qsetid AND imas_library_items.libid=:libid";
+						$query .= "imas_library_items.qsetid=:qsetid AND imas_library_items.libid=:libid ";
+						$query .= "AND imas_library_items.deleted=0";
 						$stm = $DBH->prepare($query);
 						$stm->execute(array(':groupid'=>$groupid, ':qsetid'=>$qsetid, ':libid'=>$lib));
 						if ($stm->rowCount()>0) {
@@ -209,7 +210,7 @@ if ($myrights<20) {
 						//DB $query = "SELECT id FROM imas_library_items WHERE qsetid='$qsetid'";
 						//DB $result = mysql_query($query) or die("Query failed : $query " . mysql_error());
 						//DB if (mysql_num_rows($result)==0) {
-						$stm = $DBH->prepare("SELECT id FROM imas_library_items WHERE qsetid=:qsetid");
+						$stm = $DBH->prepare("SELECT id FROM imas_library_items WHERE qsetid=:qsetid AND deleted=0");
 						$stm->execute(array(':qsetid'=>$qsetid));
 						if ($stm->rowCount()==0) {
 							//DB $query = "INSERT INTO imas_library_items (qsetid,libid,ownerid) VALUES ('$qsetid',0,$userid)";
@@ -233,7 +234,7 @@ if ($myrights<20) {
 				//DB $result = mysql_query($query) or die("Query failed : $query " . mysql_error());
 				//DB $qsetid = mysql_result($result,0,0);
 				$offset = intval($offset);
-				$stm = $DBH->prepare("SELECT qsetid FROM imas_library_items WHERE libid=:libid LIMIT $offset,1");
+				$stm = $DBH->prepare("SELECT qsetid FROM imas_library_items WHERE libid=:libid AND deleted=0 LIMIT $offset,1");
 				$stm->execute(array(':libid'=>$lib));
 				$qsetid = $stm->fetchColumn(0);
 			}
@@ -315,7 +316,8 @@ if ($myrights<20) {
 		//DB $result = mysql_query($query) or die("Query failed : $query " . mysql_error());
 		//DB $row = mysql_fetch_row($result);
 		$query = "SELECT imas_library_items.ownerid,imas_users.groupid FROM imas_library_items,imas_users WHERE ";
-		$query .= "imas_library_items.ownerid=imas_users.id AND imas_library_items.libid=:libid AND imas_library_items.qsetid=:qsetid";
+		$query .= "imas_library_items.ownerid=imas_users.id AND imas_library_items.libid=:libid AND imas_library_items.qsetid=:qsetid ";
+		$query .= "AND imas_library_items.deleted=0";
 		$stm = $DBH->prepare($query);
 		$stm->execute(array(':libid'=>$lib, ':qsetid'=>$qsetid));
 		$row = $stm->fetch(PDO::FETCH_NUM);
@@ -568,10 +570,12 @@ function delqimgs($qsid) {
 		//DB $query = "SELECT id FROM imas_qimages WHERE filename='{$row[1]}'";
 		//DB $r2 = mysql_query($query) or die("Query failed :$query " . mysql_error());
 		//DB if (mysql_num_rows($r2)==1) {
-		$stm2 = $DBH->prepare("SELECT id FROM imas_qimages WHERE filename=:filename");
-		$stm2->execute(array(':filename'=>$row[1]));
-		if ($stm2->rowCount()==1) {
-			unlink(rtrim(dirname(__FILE__), '/\\') .'/../assessment/qimages/'.$row[1]);
+		if (substr($row[1],0,4)!='http') {
+			$stm2 = $DBH->prepare("SELECT id FROM imas_qimages WHERE filename=:filename");
+			$stm2->execute(array(':filename'=>$row[1]));
+			if ($stm2->rowCount()==1) {
+				unlink(rtrim(dirname(__FILE__), '/\\') .'/../assessment/qimages/'.$row[1]);
+			}
 		}
 		//DB $query = "DELETE FROM imas_qimages WHERE id='{$row[0]}'";
 		//DB mysql_query($query) or die("Query failed :$query " . mysql_error());
